Rename SanityIT to ClientUrlIT
